refact: type fold structure
