+++ dashboard/src/features/users/components/dialogs/settings/user-info/index.ts
@@ -1,3 +1,4 @@
 export * from './table'
+export * from './subscription-actions'
 export * from './user-status-enable-button'
 export * from './qrcode'
