+++ dashboard/src/features/users/types/data-limit-reset-strategy.ts
@@ -1,0 +1,8 @@
+
+export type DataLimitResetStrategy =
+    | 'no_reset'
+    | 'day'
+    | 'week'
+    | 'month'
+    | 'year';
+
