+++ dashboard/src/features/users/types/expire-strategy.ts
@@ -1,0 +1,1 @@
+export type ExpireStrategy = "never" | "fixed_date" | "start_on_first_use";
