+++ dashboard/src/features/users/types/user-prop.ts
@@ -1,0 +1,5 @@
+import { UserType } from "@marzneshin/features/users";
+
+export interface UserProp {
+    user: UserType;
+}
